+++ src/main/java/com/telepathicgrunt/blame/main/StructureFeatureBlame.java
@@ -1,11 +1,15 @@
 package com.telepathicgrunt.blame.main;
 
 import com.telepathicgrunt.blame.Blame;
+import net.minecraft.structure.Structure;
 import net.minecraft.util.Identifier;
 import net.minecraft.util.registry.Registry;
+import net.minecraft.util.registry.RegistryKey;
 import net.minecraft.world.gen.chunk.StructureConfig;
 import net.minecraft.world.gen.feature.StructureFeature;
 import org.apache.logging.log4j.Level;
+
+import java.util.Map;
 
 /* @author - TelepathicGrunt
  *
@@ -31,5 +35,17 @@ public class StructureFeatureBlame {
                 "\n If you cannot find or fix the spacing/separation with the config, please let " +
                 "\n the mod author or datapack dev know about this so they can fix this.\n";
         Blame.LOGGER.log(Level.ERROR, errorReport);
+    }
+
+    public static void verifyStructuresInRegistry(){
+        for(Map.Entry<RegistryKey<StructureFeature<?>>, StructureFeature<?>> structure : Registry.STRUCTURE_FEATURE.getEntries()){
+            if(!StructureFeature.STRUCTURES.containsValue(structure.getValue())){
+                String errorReport = "\n****************** Blame Report Processor " + Blame.VERSION + " ******************" +
+                        "\n\n Found a structure not registered to the \"StructureFeature.STRUCTURES\" field (Mojmap)." +
+                        "\n Modders, please add your structure to that map or else chunks cannot be saved in game anymore." +
+                        "\n The missing structure from the map is: " + structure.getKey();
+                Blame.LOGGER.log(Level.ERROR, errorReport);
+            }
+        }
     }
 }