+++ src/main/java/com/telepathicgrunt/blame/mixin/CommandManagerMixin.java
@@ -35,15 +35,15 @@ public class CommandManagerMixin {
 
     @Inject(method = "<init>",
             at = @At(value = "RETURN"))
-    private void onInit(CommandManager.RegistrationEnvironment environment, CallbackInfo ci) {
+    private void blame_onInit(CommandManager.RegistrationEnvironment environment, CallbackInfo ci) {
         BrokenCommandBlame.detectBrokenCommand(dispatcher);
     }
 
     @Inject(method = "execute(Lnet/minecraft/server/command/ServerCommandSource;Ljava/lang/String;)I",
             at = @At(value = "INVOKE", target = "Lorg/apache/logging/log4j/Logger;isDebugEnabled()Z", remap = false),
             locals = LocalCapture.CAPTURE_FAILHARD)
-    private void printFailedCommandStacktrace(ServerCommandSource commandSource, String commandString, CallbackInfoReturnable<Integer> cir,
-                                              Exception exception, MutableText mutableText) {
+    private void blame_printFailedCommandStacktrace(ServerCommandSource commandSource, String commandString, CallbackInfoReturnable<Integer> cir,
+                                                    Exception exception, MutableText mutableText) {
         if (!LOGGER.isDebugEnabled()) {
             BrokenCommandBlame.printStacktrace(commandString, LOGGER, exception, mutableText);
         }