+++ src/main/java/com/telepathicgrunt/blame/mixin/SinglePoolElementMixin.java
@@ -21,7 +21,7 @@ public class SinglePoolElementMixin {
 
     @Inject(method = "method_27233(Lnet/minecraft/structure/StructureManager;)Lnet/minecraft/structure/Structure;",
             at = @At(value = "HEAD"))
-    private void storeCurrentPool(StructureManager structureManager, CallbackInfoReturnable<BlockBox> cir) {
+    private void blame_storeCurrentPool(StructureManager structureManager, CallbackInfoReturnable<BlockBox> cir) {
         if (MissingNBTBlame.CALLING_POOL != null & ((SinglePoolElementAccessor) this).blame_getTemplateID().left().isPresent()) {
             MissingNBTBlame.storeCurrentIdentifiers(new Pair<>(MissingNBTBlame.CALLING_POOL, ((SinglePoolElementAccessor) this).blame_getTemplateID().left().get()));
         }