+++ src/main/java/com/telepathicgrunt/blame/mixin/StructureFeatureMixin.java
@@ -21,7 +21,7 @@ public class StructureFeatureMixin {
 
     @Inject(method = "getStartChunk(Lnet/minecraft/world/gen/chunk/StructureConfig;JLnet/minecraft/world/gen/ChunkRandom;II)Lnet/minecraft/util/math/ChunkPos;",
             at = @At(value = "HEAD"))
-    private void checkSpacing(StructureConfig config, long worldSeed, ChunkRandom placementRandom, int chunkX, int chunkY, CallbackInfoReturnable<ChunkPos> cir) {
+    private void blame_checkSpacing(StructureConfig config, long worldSeed, ChunkRandom placementRandom, int chunkX, int chunkY, CallbackInfoReturnable<ChunkPos> cir) {
         if (config.getSpacing() == 0 || config.getSpacing() - config.getSeparation() <= 0) {
             StructureFeatureBlame.printStructureSpacingBlame((StructureFeature<?>) (Object) this, config);
         }