+++ src/main/java/com/telepathicgrunt/blame/mixin/StructurePoolBasedGeneratorMixin.java
@@ -35,11 +35,11 @@ public class StructurePoolBasedGeneratorMixin {
     @Inject(method = "method_30419(Lnet/minecraft/util/registry/DynamicRegistryManager;Lnet/minecraft/world/gen/feature/StructurePoolFeatureConfig;Lnet/minecraft/structure/pool/StructurePoolBasedGenerator$PieceFactory;Lnet/minecraft/world/gen/chunk/ChunkGenerator;Lnet/minecraft/structure/StructureManager;Lnet/minecraft/util/math/BlockPos;Lnet/minecraft/structure/StructurePiecesHolder;Ljava/util/Random;ZZLnet/minecraft/world/HeightLimitView;)V",
             at = @At(value = "INVOKE_ASSIGN", target = "Lnet/minecraft/structure/pool/StructurePool;getRandomElement(Ljava/util/Random;)Lnet/minecraft/structure/pool/StructurePoolElement;"),
             locals = LocalCapture.CAPTURE_FAILHARD)
-    private static void storeCurrentPool(DynamicRegistryManager dynamicRegistryManager, StructurePoolFeatureConfig structurePoolFeatureConfig,
-                                         StructurePoolBasedGenerator.PieceFactory pieceFactory, ChunkGenerator chunkGenerator,
-                                         StructureManager structureManager, BlockPos blockPos, StructurePiecesHolder structurePiecesHolder,
-                                         Random random, boolean bl, boolean bl2, HeightLimitView heightLimitView, CallbackInfo ci, List<PoolStructurePiece> list,
-                                         Registry<StructurePool> mutableRegistry, BlockRotation blockRotation, StructurePool structurePool) {
+    private static void blame_storeCurrentPool(DynamicRegistryManager dynamicRegistryManager, StructurePoolFeatureConfig structurePoolFeatureConfig,
+                                               StructurePoolBasedGenerator.PieceFactory pieceFactory, ChunkGenerator chunkGenerator,
+                                               StructureManager structureManager, BlockPos blockPos, StructurePiecesHolder structurePiecesHolder,
+                                               Random random, boolean bl, boolean bl2, HeightLimitView heightLimitView, CallbackInfo ci, List<PoolStructurePiece> list,
+                                               Registry<StructurePool> mutableRegistry, BlockRotation blockRotation, StructurePool structurePool) {
         MissingNBTBlame.CALLING_POOL = structurePool.getId();
     }
 }